requirements.txt
Django==1.11.21
djangorestframework==3.5.3
